Move PreferredProperties.PartitioningProperties to PreferredPartitioning
